refactor(claude): extract session types and utils into separate files
